+++ Frameworks/CleverPush.xcframework/ios-arm64_x86_64-simulator/CleverPush.framework/Headers/CleverPushInstance.h
@@ -18,6 +18,7 @@
 #import "CPChannelTag.h"
 #import "CPChannelTopic.h"
 #import "CPIabTcfMode.h"
+#import "CPGroupNotificationSoundMode.h"
 
 @interface CPNotificationReceivedResult : NSObject
 
@@ -193,12 +194,14 @@ extern NSString* _Nullable const CLEVERPUSH_SDK_VERSION;
 - (void)setAppBannerDraftsEnabled:(BOOL)showDraft;
 - (void)setSubscriptionChanged:(BOOL)subscriptionChanged;
 - (void)setIncrementBadge:(BOOL)increment;
+- (void)setGroupNotificationSoundMode:(CPGroupNotificationSoundMode)mode;
 - (void)setShowNotificationsInForeground:(BOOL)show;
 - (void)setDisplayAlertEnabledForNotifications:(BOOL)enabled;
 - (void)setSoundEnabledForNotifications:(BOOL)enabled;
 - (void)setBadgeCountEnabledForNotifications:(BOOL)enabled;
 - (void)setIgnoreDisabledNotificationPermission:(BOOL)ignore;
 - (void)setAutoRequestNotificationPermission:(BOOL)autoRequest;
+- (void)setProvisionalNotificationAuthorizationEnabled:(BOOL)enabled;
 - (void)setKeepTargetingDataOnUnsubscribe:(BOOL)keepData;
 - (void)addChatView:(CPChatView* _Nullable)chatView;
 - (void)showTopicsDialog;
@@ -245,6 +248,8 @@ extern NSString* _Nullable const CLEVERPUSH_SDK_VERSION;
 - (void)removeAllNotifications;
 - (void)removeNotification:(NSString* _Nullable)notificationId removeFromNotificationCenter:(BOOL)removeFromCenter;
 - (void)trackInboxClicked:(NSString* _Nullable)notificationId;
+- (void)trackNotificationDelivered:(NSString* _Nonnull)notificationId;
+- (void)trackNotificationClicked:(NSString* _Nonnull)notificationId;
 - (void)setNotificationRead:(NSString* _Nullable)notificationId read:(BOOL)read;
 - (BOOL)getNotificationRead:(NSString* _Nullable)notificationId;
 - (void)setMaximumNotificationCount:(int)limit;
@@ -267,6 +272,7 @@ extern NSString* _Nullable const CLEVERPUSH_SDK_VERSION;
 - (int)getLocalEventTrackingRetentionDays;
 - (void)getBadgeCount:(void (^ _Nullable)(NSInteger))completionHandler;
 - (CPIabTcfMode)getIabTcfMode;
+- (CPGroupNotificationSoundMode)getGroupNotificationSoundMode;
 
 - (UIColor* _Nullable)getBrandingColor;